city_trend_dashboard/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pytrends
pandas
